fixing delete error

+++ backend/product/views.py
@@ -127,10 +127,12 @@ class DeleteAllProductsView(generics.GenericAPIView):
     permission_classes = [IsAuthenticated]
 
     def delete(self, request, *args, **kwargs):
-        batch_size = 5
+        batch_size = 10
         qs = Products.objects.all()
         while qs.exists():
             ids = list(qs.values_list("id", flat=True)[:batch_size])
+            if not ids:
+                break
             Products.objects.filter(id__in=ids).delete()
         return custom_response(data=None, method="DELETE", data_name="Products")
